Abstract user-facing public API for ClientResponse
    It's important to notice that the __init__ method varies between
    implementations and there are some leftovers (e.g, the _in_context field).
    The base class is basically an interface.
